Fixed a signal repaint bug when showing measurement view icw swapped channels.

+++ src/main/java/nl/lxtreme/test/dnd/CursorTransferable.java
@@ -69,11 +69,11 @@ public class CursorTransferable implements Transferable
   /**
    * Returns the index of the cursor denoted by this transferable.
    * 
-   * @return the cursor index, as object, never <code>null</code>.
+   * @return the cursor index, >= 0.
    */
-  public Integer getCursorIdx()
+  public int getCursorIdx()
   {
-    return this.cursorIdx;
+    return this.cursorIdx.intValue();
   }
 
   /**
